requirements.txt
Flask==1.1.1
python-dotenv==0.13.0
requests==2.22.0
tweepy==3.8.0
